Removed unused class TestListener
