+++ src/test/java/tests/BaseTest.java
@@ -9,6 +9,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import pages.*;
 import utils.CapabilitiesGenerator;
+import utils.TestListener;
 
 import java.util.concurrent.TimeUnit;
