+++ src/test/java/utils/TestListener.java
@@ -1,51 +1,51 @@
-//package utils;
-//
-//import io.qameta.allure.Attachment;
-//import org.openqa.selenium.NoSuchSessionException;
-//import org.openqa.selenium.OutputType;
-//import org.openqa.selenium.TakesScreenshot;
-//import org.openqa.selenium.WebDriver;
-//import org.testng.ITestListener;
-//import org.testng.ITestResult;
-//
-//import java.util.concurrent.TimeUnit;
-//
-//
-//public class TestListener implements ITestListener {
-//
-//    public void onTestStart(ITestResult iTestResult) {
-//        System.out.println(String.format("======================================== STARTING TEST %s ========================================", iTestResult.getName()));
-//    }
-//
-//    public void onTestSuccess(ITestResult iTestResult) {
-//        System.out.println(String.format("======================================== FINISHED TEST %s Duration: %ss ========================================", iTestResult.getName(),
-//                getExecutionTime(iTestResult)));
-//    }
-//
-//    public void onTestFailure(ITestResult iTestResult) {
-//        System.out.println(String.format("======================================== FAILED TEST %s Duration: %ss ========================================", iTestResult.getName(),
-//                getExecutionTime(iTestResult)));
-//        takeScreenshot(iTestResult);
-//    }
-//
-//    public void onTestSkipped(ITestResult iTestResult) {
-//        System.out.println(String.format("======================================== SKIPPING TEST %s ========================================", iTestResult.getName()));
-//        takeScreenshot(iTestResult);
-//    }
-//
-//    @Attachment(value = "Last screen state", type = "image/png")
-//    private byte[] takeScreenshot(ITestResult iTestResult) {
-//        try {
-//            return ((TakesScreenshot) getWebDriver()).getScreenshotAs(OutputType.BYTES);
-//        } catch (NoSuchSessionException ex) {
-//            return null;
-//        } catch (IllegalStateException ex) {
-//            return null;
-//        }
-//    }
-//
-//
-//    private long getExecutionTime(ITestResult iTestResult) {
-//        return TimeUnit.MILLISECONDS.toSeconds(iTestResult.getEndMillis() - iTestResult.getStartMillis());
-//    }
-//}
+package utils;
+
+import io.qameta.allure.Attachment;
+import org.openqa.selenium.NoSuchSessionException;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.WebDriver;
+import org.testng.ITestContext;
+import org.testng.ITestListener;
+import org.testng.ITestResult;
+
+import java.util.concurrent.TimeUnit;
+
+public class TestListener implements ITestListener {
+    public void onTestStart(ITestResult iTestResult) {
+        System.out.println(String.format("======================================== STARTING TEST %s ========================================", iTestResult.getName()));
+    }
+
+    public void onTestSuccess(ITestResult iTestResult) {
+        System.out.println(String.format("======================================== FINISHED TEST %s Duration: %ss ========================================", iTestResult.getName(),
+                getExecutionTime(iTestResult)));
+    }
+
+    public void onTestFailure(ITestResult iTestResult) {
+        System.out.println(String.format("======================================== FAILED TEST %s Duration: %ss ========================================", iTestResult.getName(),
+                getExecutionTime(iTestResult)));
+        takeScreenshot(iTestResult);
+    }
+
+    public void onTestSkipped(ITestResult iTestResult) {
+        System.out.println(String.format("======================================== SKIPPING TEST %s ========================================", iTestResult.getName()));
+        takeScreenshot(iTestResult);
+    }
+
+    @Attachment(value = "Last screen state", type = "image/png")
+    private byte[] takeScreenshot(ITestResult iTestResult) {
+        try {
+            ITestContext context = iTestResult.getTestContext();
+            WebDriver driver = (WebDriver) context.getAttribute("driver");
+            return ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
+        } catch (NoSuchSessionException ex) {
+            return null;
+        } catch (IllegalStateException ex) {
+            return null;
+        }
+    }
+
+    private long getExecutionTime(ITestResult iTestResult) {
+        return TimeUnit.MILLISECONDS.toSeconds(iTestResult.getEndMillis() - iTestResult.getStartMillis());
+    }
+}
